requirements.txt
rdkit-pypi==2022.9.5